fix(admin): split the Shell nav fork and baseline the wash specimen for CI
CI's source-structure gate caps new files at 350 lines; the forked
Shell/NavigationBar.tsx landed at 490. The FAB + speed dial move to their own
Shell/FabButton.tsx (264 lines; NavigationBar drops to 236) with a sibling
story, behavior unchanged. The tokens-gallery story line documenting the 1a
canvas wash gradient gets its audited storybook-fixture baseline entry so the
design-tokens gate passes on the line it was written to document.
