+++ packages/admin/src/components/Shell/index.ts
@@ -1,3 +1,4 @@
 export { AppBar, type AppBarProps } from "./AppBar";
 export { MainSheet, type MainSheetProps } from "./MainSheet";
 export { NavigationBar } from "./NavigationBar";
+export { FabButton, type FabButtonProps } from "./FabButton";
