+++ scripts/data/design-token-usage-baseline.tsv
@@ -132,3 +132,4 @@ packages/shared/src/components/Dialog/ImagePreviewDialog.tsx <span className="ml
 packages/shared/src/components/Dialog/ImagePreviewDialog.tsx className="btn-icon bg-bg-white-0/10 tap-feedback text-white rounded-full sm:ml-2"	legacy-runtime	shared	2026-12-31	white-on-image overlay chrome; tokens defer to image-preview palette work
 packages/shared/src/components/Dialog/ImagePreviewDialog.tsx className="btn-icon bg-bg-white-0/20 hover:bg-bg-white-0/30 tap-feedback text-white rounded-full"	legacy-runtime	shared	2026-12-31	white-on-image overlay chrome; tokens defer to image-preview palette work
 packages/shared/src/components/Dialog/ImagePreviewDialog.tsx className="hidden sm:flex btn-icon bg-bg-white-0/10 tap-feedback text-white rounded-full"	legacy-runtime	shared	2026-12-31	white-on-image overlay chrome; tokens defer to image-preview palette work
+packages/admin/src/components/AdminRevampTokens.stories.tsx "linear-gradient(180deg, var(--tone-surface-tint-color, transparent), transparent 320px), rgb(var(--m3-surface-container-low))",	storybook-fixture	admin	2026-12-31	token-gallery specimen documents the 1a canvas wash gradient
